# backend/users/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "User Microservice for Console d",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/user/v1/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new user with the provided user data.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Creates a new user",
                "parameters": [
                    {
                        "description": "User object containing user data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.User"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/user/v1/delete/{id}/{hardDelete}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a user with the provided user id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Deletes a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Hard delete or soft delete",
                        "name": "hardDelete",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/user/v1/fieldCatalogues": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Gets all fields catalogues of user service tables.",
                "produces": [
                    "application/json"
                ],
                "summary": "Fields Catalogues of user service",
                "responses": {}
            }
        },
        "/api/user/v1/get": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get or Read users from database based on filters if any.",
                "produces": [
                    "application/json"
                ],
                "summary": "Get or Read users from database based on filters if any",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filters for user retrieval",
                        "name": "filters",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/api/user/v1/login": {
            "post": {
                "description": "Verify the User by login api with email and password.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Verify the User by login api with email and password",
                "parameters": [
                    {
                        "description": "Object containing email, password and token",
                        "name": "loginCrds",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.LoginCredentials"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/user/v1/register": {
            "post": {
                "description": "Register a new user with the provided user data.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "User object containing user data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.User"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/user/v1/sendOTPVerificationMail": {
            "post": {
                "description": "Send OTP Verification Mail To User.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Send OTP Verification Mail To User",
                "parameters": [
                    {
                        "description": "Object containing email",
                        "name": "email",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.SendOTPVerificationMail"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/user/v1/update/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates a user with the provided user data.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Updates a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User object containing user data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.User"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/user/v1/updatePassword/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates a user password with the provided user password data.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Updates a user password",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User object containing user password data",
                        "name": "password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Password"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/user/v1/verifyOTP": {
            "post": {
                "description": "Verify the OTP from User Mail.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Verify the OTP from User Mail",
                "parameters": [
                    {
                        "description": "Object containing otp and email",
                        "name": "otp",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.VerifyOTP"
                        }
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "main.Address": {
            "type": "object",
            "properties": {
                "country": {
                    "type": "string"
                },
                "district": {
                    "type": "string"
                },
                "doorNo": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "pinCode": {
                    "type": "integer"
                },
                "state": {
                    "type": "string"
                },
                "street": {
                    "type": "string"
                },
                "village": {
                    "type": "string"
                }
            }
        },
        "main.LoginCredentials": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "refreshToken": {
                    "type": "string"
                }
            }
        },
        "main.Organization": {
            "type": "object",
            "properties": {
                "contactEmail": {
                    "type": "string"
                },
                "contactPerson": {
                    "type": "string"
                },
                "contactPhoneNumber": {
                    "type": "string"
                },
                "domain": {
                    "type": "string"
                },
                "industry": {
                    "type": "string"
                },
                "organizationName": {
                    "type": "string"
                }
            }
        },
        "main.Password": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "main.Personal": {
            "type": "object",
            "properties": {
                "dateOfBirth": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "integer"
                }
            }
        },
        "main.SendOTPVerificationMail": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "main.Settings": {
            "type": "object",
            "properties": {
                "currencyCode": {
                    "type": "string"
                },
                "dateFormat": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                }
            }
        },
        "main.User": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password",
                "region",
                "userType"
            ],
            "properties": {
                "accountNumber": {
                    "type": "integer"
                },
                "active": {
                    "type": "boolean"
                },
                "address": {
                    "$ref": "#/definitions/main.Address"
                },
                "cloudProvider": {
                    "type": "string"
                },
                "createdOn": {
                    "type": "string"
                },
                "deleted": {
                    "type": "boolean"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "object",
                    "additionalProperties": true
                },
                "lastChangedBy": {
                    "type": "string"
                },
                "lastChangedOn": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "organization": {
                    "$ref": "#/definitions/main.Organization"
                },
                "password": {
                    "type": "string"
                },
                "personal": {
                    "$ref": "#/definitions/main.Personal"
                },
                "region": {
                    "type": "string"
                },
                "saltStored": {
                    "type": "string"
                },
                "settings": {
                    "$ref": "#/definitions/main.Settings"
                },
                "userType": {
                    "type": "string"
                }
            }
        },
        "main.VerifyOTP": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "otp": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer token authentication",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}